+++ packages/jest-cli/src/args.ts
@@ -172,7 +172,7 @@ export const options: {[key: string]: Options} = {
     requiresArg: true,
     type: 'string',
   },
-  collectOnly: {
+  collectTests: {
     description:
       'Discover and report all test cases without executing them. ' +
       'Prints a tree of test suites and test names.',