+++ packages/jest-config/src/normalize.ts
@@ -894,7 +894,7 @@ export default async function normalize(
       case 'changedFilesWithAncestor':
       case 'clearMocks':
       case 'collectCoverage':
-      case 'collectOnly':
+      case 'collectTests':
       case 'coverageProvider':
       case 'coverageReporters':
       case 'coverageThreshold':