+++ packages/jest-schemas/src/raw-types.ts
@@ -237,7 +237,7 @@ export const InitialOptions = Type.Partial(
     changedSince: Type.String(),
     collectCoverage: Type.Boolean(),
     collectCoverageFrom: Type.Array(Type.String()),
-    collectOnly: Type.Boolean(),
+    collectTests: Type.Boolean(),
     coverageDirectory: Type.String(),
     coveragePathIgnorePatterns: Type.Array(Type.String()),
     coverageProvider: CoverageProvider,