+++ packages/jest-types/src/Config.ts
@@ -261,7 +261,7 @@ export type GlobalConfig = {
   ci: boolean;
   collectCoverage: boolean;
   collectCoverageFrom: Array<string>;
-  collectOnly: boolean;
+  collectTests: boolean;
   coverageDirectory: string;
   coveragePathIgnorePatterns?: Array<string>;
   coverageProvider: CoverageProvider;
@@ -413,7 +413,7 @@ export type Argv = Arguments<
     clearMocks: boolean;
     collectCoverage: boolean;
     collectCoverageFrom: string;
-    collectOnly: boolean;
+    collectTests: boolean;
     color: boolean;
     colors: boolean;
     config: string;